Move setup out of timed task
